requirements.txt
python-dotenv==0.10.3
requests==2.22.0
urllib3==1.25.7
pytest==5.3.2
